feat/runtime changes (#35)
* Moved things around, added scene object, added basic ability to load in body during runtime

* improved sphere rendering (again), removed redundant code

* Removed unused lines. Legend is now broken for some reason?

* WIP: Changing definition of Matrix4 and adding rotation to planets

* Basic Matrix4 changes

* Added more Matrix4 funcs, replaced legacy matrix defs with new ones

* decoupled physics stepping from the render loop (#34)

ref: https://gafferongames.com/post/fix_your_timestep/

- added accumulator loop to both 2D and 3D rander loops

- enabled vsync on both renderers

- bump 3D camera rotation speed

- added clang formatting


* fixed physics / render decouple, found bug in perspective matrix

* Fixed perspective matrix bug, added MATRIX4_INIT_EMPTY, found bug I can't fix in decoupling

* fixed that last one. finally got an approach I like for time here

* Added rotation to bodies, tied to time in si

* Ran clang-format on everything, with the included format file

* updated readme, trying to fix broken legend

* Fixed legend bug

* Fixed bugs in the runtime addition function

* fixing small mistake I found

* small fixes. still investigating bug in physics pipeline

* adding simulation mode enum

* Added simulation settings. Think I ironed out DT bug

---------

+++ include/math/matrix/matrix3.h
@@ -1,7 +1,7 @@
 #ifndef MATRIX3_H
 #define MATRIX3_H
 
-//3x3 matrix represented as a 2d array
-typedef double matrix3 [3][3];
+// 3x3 matrix represented as a 2d array
+typedef float matrix3[3][3];
 
 #endif
